arch/mips64r6el: introduce MIPS64R6EL_USE_MSA (enabled by default)
Introduce a switch for MSA flags. CIP does not have a clear guideline on SIMD
support on their products yet (they will be the sole IP designer for MIPS r6
cores). There is a P6600 core which has optional MSA, while I6400/6500/6600
come with MSA as standard. P6600 is not expected to see production.

This branch will be used for our experimental MSA port, made available to
CIP's hardware departments for evaluation purposes.